Refactor worker to add testMode and usage info
Needed to test reaping functionality independently from a full working
worker
