+++ pancancer-arch-3/scripts/run_worker_daemon.sh
@@ -9,8 +9,7 @@ if [ -f $PID_FILE ]; then
   exit 0
 fi
 set -e
-#Note: -Dpidfile is not a standard java option; it is custom, for the Worker.
-nohup java -Dpidfile=$PID_FILE -cp pancancer-arch-3-1.0.0-SNAPSHOT.jar info.pancancer.arch3.worker.Worker --config workerConfig.json --uuid 50f20496-c221-4c25-b09b-839511e76df4 </dev/null > $LOG_FILE 2>&1 &
+nohup java -cp pancancer-arch-3-1.0.0-SNAPSHOT.jar info.pancancer.arch3.worker.Worker --config workerConfig.json --uuid 50f20496-c221-4c25-b09b-839511e76df4 --pidFile $PID_FILE  </dev/null > $LOG_FILE 2>&1 &
 PID=$!
 echo $PID > $PID_FILE
 echo "PID of worker daemon is $PID"
