+++ pancancer-arch-3/src/test/java/info/pancancer/arch3/test/TestWorker.java
@@ -1,15 +1,18 @@
 package info.pancancer.arch3.test;
 
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
-import static org.mockito.Matchers.any;
-import static org.mockito.Matchers.anyString;
+import com.rabbitmq.client.AMQP.BasicProperties;
+import com.rabbitmq.client.Channel;
+import com.rabbitmq.client.ConsumerCancelledException;
+import com.rabbitmq.client.Envelope;
+import com.rabbitmq.client.QueueingConsumer;
+import com.rabbitmq.client.QueueingConsumer.Delivery;
+import com.rabbitmq.client.ShutdownSignalException;
 import info.pancancer.arch3.beans.Job;
 import info.pancancer.arch3.utils.Utilities;
 import info.pancancer.arch3.worker.Worker;
 import info.pancancer.arch3.worker.WorkerHeartbeat;
+import info.pancancer.arch3.worker.WorkerRunnable;
 import info.pancancer.arch3.worker.WorkflowRunner;
-
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
@@ -18,12 +21,14 @@ import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.concurrent.ExecutionException;
-
 import org.json.simple.JSONObject;
+import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.fail;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import static org.mockito.Matchers.any;
+import static org.mockito.Matchers.anyString;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
@@ -35,21 +40,11 @@ import org.powermock.modules.junit4.PowerMockRunner;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.rabbitmq.client.AMQP.BasicProperties;
-import com.rabbitmq.client.Channel;
-import com.rabbitmq.client.ConsumerCancelledException;
-import com.rabbitmq.client.Envelope;
-import com.rabbitmq.client.QueueingConsumer;
-import com.rabbitmq.client.QueueingConsumer.Delivery;
-import com.rabbitmq.client.ShutdownSignalException;
-
-@PrepareForTest({ QueueingConsumer.class, Utilities.class, Worker.class, WorkerHeartbeat.class, WorkflowRunner.class, Logger.class,
-        LoggerFactory.class })
+@PrepareForTest({ QueueingConsumer.class, Worker.class, WorkerRunnable.class, Utilities.class, WorkerHeartbeat.class, WorkflowRunner.class,
+        Logger.class, LoggerFactory.class })
 @RunWith(PowerMockRunner.class)
 public class TestWorker {
 
-    private static final Logger log = LoggerFactory.getLogger(TestWorker.class);
-    
     @Mock
     private WorkerHeartbeat mockHeartbeat;
 
@@ -80,14 +75,12 @@ public class TestWorker {
         @Override
         public Object answer(InvocationOnMock invocation) throws Throwable {
             outBuffer.append(invocation.getArguments()[0]).append("\n");
-            if (invocation.getArguments().length == 2)
-            {
-                if (invocation.getArguments()[1] instanceof Throwable)
-                {
+            if (invocation.getArguments().length == 2) {
+                if (invocation.getArguments()[1] instanceof Throwable) {
                     OutputStream outStream = new ByteArrayOutputStream();
                     PrintStream ps = new PrintStream(outStream);
-                    ((Throwable)invocation.getArguments()[1]).printStackTrace(ps);
-                    outBuffer.append(new String (outStream.toString()));
+                    ((Throwable) invocation.getArguments()[1]).printStackTrace(ps);
+                    outBuffer.append(new String(outStream.toString()));
                     ps.close();
                     outStream.close();
                 }
@@ -101,7 +94,7 @@ public class TestWorker {
     @Before
     public void setup() throws IOException, Exception {
         MockitoAnnotations.initMocks(this);
-        
+
         outBuffer = new StringBuffer();
         Mockito.doAnswer(logAnswer).when(mockLogger).info(anyString());
         Mockito.doAnswer(logAnswer).when(mockLogger).debug(anyString());
@@ -109,6 +102,7 @@ public class TestWorker {
         Mockito.doAnswer(logAnswer).when(mockLogger).error(anyString(), any(Exception.class));
         PowerMockito.mockStatic(org.slf4j.LoggerFactory.class);
         Mockito.when(org.slf4j.LoggerFactory.getLogger(Worker.class)).thenReturn(mockLogger);
+        Mockito.when(org.slf4j.LoggerFactory.getLogger(WorkerRunnable.class)).thenReturn(mockLogger);
 
         PowerMockito.mockStatic(Utilities.class);
 
@@ -132,7 +126,7 @@ public class TestWorker {
         PowerMockito.mockStatic(Utilities.class);
         Mockito.when(Utilities.parseConfig(anyString())).thenReturn(new JSONObject());
         try {
-            Worker testWorker = new Worker("src/test/resources/workerConfig.json", "vm123456", 1);
+            WorkerRunnable testWorker = new WorkerRunnable("src/test/resources/workerConfig.json", "vm123456", 1);
             fail("Execution should not have reached this point!");
         } catch (Exception e) {
             assertTrue(
@@ -154,7 +148,7 @@ public class TestWorker {
         Delivery testDelivery = new Delivery(mockEnvelope, mockProperties, body);
         setupMockQueue(testDelivery);
 
-        Worker testWorker = new Worker("src/test/resources/workerConfig.json", "vm123456", 1);
+        WorkerRunnable testWorker = new WorkerRunnable("src/test/resources/workerConfig.json", "vm123456", 1);
 
         testWorker.run();
         String testResults = outBuffer.toString();
@@ -169,7 +163,7 @@ public class TestWorker {
         Delivery testDelivery = new Delivery(mockEnvelope, mockProperties, body);
         setupMockQueue(testDelivery);
 
-        Worker testWorker = new Worker("src/test/resources/workerConfig.json", "vm123456", 1);
+        WorkerRunnable testWorker = new WorkerRunnable("src/test/resources/workerConfig.json", "vm123456", 1);
         try {
             testWorker.run();
             String testResults = outBuffer.toString();
@@ -187,7 +181,7 @@ public class TestWorker {
         Delivery testDelivery = new Delivery(mockEnvelope, mockProperties, null);
         setupMockQueue(testDelivery);
 
-        Worker testWorker = new Worker("src/test/resources/workerConfig.json", "vm123456", 1);
+        WorkerRunnable testWorker = new WorkerRunnable("src/test/resources/workerConfig.json", "vm123456", 1);
         try {
             testWorker.run();
             String testResults = outBuffer.toString();
@@ -208,7 +202,7 @@ public class TestWorker {
         Worker.main(new String[] { "--config", "src/test/resources/workerConfig.json", "--uuid", "vm123456", "--max-runs", "1" });
 
         String testResults = outBuffer.toString();
-        
+
         testResults = cleanResults(testResults);
 
         String begining = new String(Files.readAllBytes(Paths.get("src/test/resources/testResult_Start.txt")));
@@ -252,7 +246,7 @@ public class TestWorker {
         testResults = cleanResults(testResults);
 
         String begining = new String(Files.readAllBytes(Paths.get("src/test/resources/testResult_Start.txt")));
-        //System.out.println(testResults);
+        // System.out.println(testResults);
         assertTrue("check begining of output", testResults.contains(begining));
 
         String ending = new String(Files.readAllBytes(Paths.get("src/test/resources/testResult_End.txt")));
@@ -269,7 +263,7 @@ public class TestWorker {
         Delivery testDelivery = new Delivery(mockEnvelope, mockProperties, body);
         setupMockQueue(testDelivery);
 
-        Worker testWorker = new Worker("src/test/resources/workerConfig.json", "vm123456", 1);
+        WorkerRunnable testWorker = new WorkerRunnable("src/test/resources/workerConfig.json", "vm123456", 1);
 
         testWorker.run();
 
@@ -292,7 +286,7 @@ public class TestWorker {
         j.setWorkflowVersion("1.0-SNAPSHOT");
         j.setJobHash("asdlk2390aso12jvrej");
         j.setUuid("1234567890");
-        Map<String, String> iniMap = new HashMap<String, String>(3);
+        Map<String, String> iniMap = new HashMap<>(3);
         iniMap.put("param1", "value1");
         iniMap.put("param2", "value2");
         iniMap.put("param3", "help I'm trapped in an INI file");
